mv enums -> common/
