+++ common/Cargo.toml
@@ -2,3 +2,6 @@
 name = "common"
 version = "0.2.0"
 edition = "2021"
+
+[dependencies]
+strum_macros = "0.25.3"
